initial_livechat_check/requirements.txt
beautifulsoup4==4.9.1
requests==2.21.0
python-dotenv==0.14.0
google-cloud-storage==1.29.0
google-cloud-pubsub==1.6.0
retry==0.9.2
